Work around a `pulumi plugin install` bug
When installing a plugin from Github, Pulumi currently substitutes a
version string _without_ the leading "v". Unfortunately, this results
in invalid download URLs.

To work around this centrally, we'll just add that leading "v" in
`resources.go`, rather than doing a lot of gnarly `sed`
substitutions. This also has the benefit of working for all SDKs.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>