Create `sdk/*` tags for Go SDK
When a Go module is not rooted at the base of a repository, we need a
version tag that is prefixed with the module's path in the repository.

Thus, for the Go SDK, which is found in the `sdk` directory, we need
to create tags like `sdk/vX.Y.Z` in addition to the normal `vX.Y.Z`
tags we create for a release.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>