Validate plugin installation for Go SDK
With Go, plugin installation occurs during the `pulumi` run itself,
rather than when the SDK is downloaded. As such, we'll create a
minimal Pulumi program that can trigger the download of the plugin
binaries. Then, we'll validate the binary is present and correct, as
we do for the Python and NodeJS SDKs.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>